+++ test/src/functions/vTrue.test.mjs
@@ -5,16 +5,16 @@ import assert           from 'node:assert/strict';
 import CONSTANTS        from '../../constants.mjs';
 import Validator        from '../../../src/index.mjs';
 
-describe('vIsTrue', () => {
+describe('vTrue', () => {
     it('Should be invalid if not passed a boolean', () => {
         for (const el of CONSTANTS.NOT_BOOLEAN) {
             assert.deepEqual(
-                new Validator({a: 'is_true'}).validate({a: el}),
+                new Validator({a: 'true'}).validate({a: el}),
                 {
                     is_valid: false,
                     count: 1,
                     errors: {
-                        a: [{msg: el === undefined ? 'not_found' : 'is_true', params: []}],
+                        a: [{msg: el === undefined ? 'not_found' : 'true', params: []}],
                     },
                 }
             );
@@ -23,14 +23,14 @@ describe('vIsTrue', () => {
 
     it('Should be invalid if passed false', () => {
         assert.deepEqual(
-            new Validator({a: 'is_true'}).validate({a: false}),
-            {is_valid: false, count: 1, errors: {a: [{msg: 'is_true', params: []}]}}
+            new Validator({a: 'true'}).validate({a: false}),
+            {is_valid: false, count: 1, errors: {a: [{msg: 'true', params: []}]}}
         );
     });
 
     it('Should be valid if passed true', () => {
         assert.deepEqual(
-            new Validator({a: 'is_true'}).validate({a: true}),
+            new Validator({a: 'true'}).validate({a: true}),
             {is_valid: true, count: 0, errors: {}}
         );
     });